+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2020-9-5
+### Added
+- Item picker dialog can be closed using escape or movement keys
+- Keys which can close the item picker dialog are configurable
+- README.md Features and info about which keys can close the dialog
+
 ## [1.2.2] - 2020-8-29
 ### Fixed
 - README.md for thunderstore.io